LayoutTests/crypto/crypto-key-algorithm-gc-expected.txt

Test that CryptoKey.algorithm preserves custom properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Generating a HMAC key with default length...
PASS key.algorithm === key.algorithm is true
PASS key.algorithm.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
